Move impls to ssh/mod.rs
